refactor: extract WeekPage and WeekBody from week_view
